requirements.txt
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dacite==1.9.2
filetype==1.2.0
fonttools==4.60.0
idna>=2.5,<4
ImageHash==4.3.1
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
minify_html==0.16.4
multimethod==1.12
networkx==3.5
numba>=0.61.0,<0.62.0
numpy>=1.25,<2.0
packaging==25.0
pandas>=2.3.2,<2.4
patsy==1.0.1
phik==0.12.5
pillow==11.3.0
puremagic==1.30
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets>=1.9.0,<2.0
PyYAML==6.0.2
requests==2.32.5
scipy>=1.15.3,<2.0
seaborn>=0.12.2,<0.13
six==1.17.0
statsmodels>=0.14.5,<0.15
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
visions==0.8.1
wordcloud>=1.9.4,<2.0
ydata-profiling>=4.17.0,<5.0
streamlit>=1.26.1,<2.0
beautifulsoup4==4.12.2
selenium==4.14.0
webdriver-manager==3.9.0
googletrans==4.0.0-rc1
deep-translator==1.11.1
langdetect==1.0.9
qrcode==7.4.2
reportlab==4.1.0
python-barcode
streamlit-aggrid
fpdf
dropbox
openpyxl
xlrd
pyarrow
gspread
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client





